+++ CHANGELOG.md
@@ -9,6 +9,7 @@ method.
 ### Fixed
 
 - Fixed security vulnerabilities by updating dependencies.
+- Suppressed annoying tokenization-related warning in `NamedEntityRecognition.__call__()` method.
 
 ## Release v0.4.0 - December 4, 2025